+++ packages/components/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @xernerx/components
 
+## 0.2.19
+
+### Patch Changes
+
+- Updated dependencies
+    - @xernerx/lib@0.5.2
+    - @xernerx/providers@0.3.12
+    - @xernerx/feedback@0.0.20
+    - @xernerx/ui@0.3.3
+
 ## 0.2.18
 
 ### Patch Changes